For readability, renamed my 'type' property in the RoomType enum to be called 'title'.  Refactored accordingly.
Added a getTitle method to both RoomType enum and again for Room (to make it simpler later on for calling this method).

Added 2 new tests to BedRoomTest class - to check we can get the whole enum type (canGetType()) for that room, and also can get the 'title' property of the associated enum type too.